separate conn from main
